requirement.txt
pandas
sqlalchemy
mysql-connector-python
pytest
